Program that prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long long int num_1 = 0, num_2 = 1, sum; 
+	long int num_1 = 0, num_2 = 1, sum; 
 	int iterate;
 
 	for (iterate = 0; iterate < 50; iterate++)
